@@ -5,9 +5,9 @@ namespace School;
 class Pupil
 {
     /**
-     * @var integer
+     * @var int
      */
-    private $id;
+    private int $id;
 
     public function __construct(int $id)
     {
@@ -16,7 +16,7 @@ class Pupil
 
 
     /**
-     * @return integer
+     * @return int
      */
     public function getId() : int
     {
